Create lexeg1.l
You first have to go to the directory in which the file lexeg1.l is in. Then using terminal/cmd do lex lexeg1.l which will make the file lex.yy.c. To the run the program you need compile it with a c compiler such as gcc. With gcc you can compile it using gcc -lfl lex.yy.c. This will create a.out which can be run using ./a.out